% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lens_info.R
\name{add_lens_data_nikon_AF_28_105_D}
\alias{add_lens_data_nikon_AF_28_105_D}
\alias{add_lens_data_nikon_AF_24_D}
\alias{add_lens_data_nikon_AF_50_D}
\alias{add_lens_data_nikon_AF_85}
\alias{add_lens_data_nikon_AF_105_D}
\alias{add_lens_data_sigma_AF_28_70_D}
\alias{delete_lens_data}
\title{Lens data helpers for common Nikon and Sigma lenses}
\usage{
add_lens_data_nikon_AF_28_105_D()

add_lens_data_nikon_AF_24_D()

add_lens_data_nikon_AF_50_D()

add_lens_data_nikon_AF_85()

add_lens_data_nikon_AF_105_D()

add_lens_data_sigma_AF_28_70_D()

delete_lens_data()
}
\value{
A one-row data.frame with Nikon lens EXIF fields:
  `Nikon:LensIDNumber`, `Nikon:LensFStops`, `Nikon:MinFocalLength`,
  `Nikon:MaxFocalLength`, `Nikon:MaxApertureAtMinFocal`,
  `Nikon:MaxApertureAtMaxFocal`, `Nikon:MCUVersion`, `Nikon:LensType`,
  `Nikon:LensSpec`, `Nikon:LensID`.
}
\description{
Each function returns a one-row data.frame of Nikon EXIF lens fields that
can be passed directly to [editing_exif()] via the `metadata` argument or
merged with an existing metadata table.
}
\details{
To help `exiftoolr` to guess the lens ID, you can provide a list of
   favourite lenses in a config file:
   # ~/.ExifTool_config
   # A special 'Lenses' list can be defined to give priority to specific lenses
   # in the logic to determine a lens model for the Composite:LensID tag
   @Image::ExifTool::UserDefined::Lenses = (
       'AF Zoom-Nikkor 28-105mm f/3.5-4.5D IF',
       'AF Nikkor 50mm f/1.4D',
       'AF Nikkor 85mm f/1.8',
       'AF Nikkor 24mm f/2.8D'
   );

  Additional lenses can be added here, the information listed here can be
  found in the exif data of pictures taken with the additional lenses.

  `Nikon:LensType` = 2 for D lenses and 0 for non-D lenses.
  Exact values for the fields such as `Nikon:MinFocalLength` were obtained
  by shooting the lens with a DSLR and looking at the EXIF data.
}
\examples{
add_lens_data_nikon_AF_28_105_D()
add_lens_data_nikon_AF_50_D()
}
\seealso{
See <https://exiftool.org/TagNames/Nikon.html#LensID> for the list
   of standard lens names.
}
