% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APMD-package.R
\docType{package}
\name{APMD-package}
\alias{APMD}
\alias{APMD-package}
\title{APMD: Analog Photography MetaData - Edit Exif Data of Pictures}
\description{
The APMD package is developed in GitHub
(https://github.com/albansagouis/APMD). To see the preferable citation of
the package, type citation("APMD").
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/albansagouis/APMD/}
  \item Report bugs at \url{https://github.com/albansagouis/APMD/issues}
}

}
\keyword{internal}
