% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_missing_records.R
\name{insert_missing_records}
\alias{insert_missing_records}
\title{Insert rows in metadata data.tables}
\usage{
insert_missing_records(
  metadata,
  row_indices,
  new_rows,
  extrapolate_data = FALSE
)
}
\arguments{
\item{metadata}{a data.frame as provided by \code{\link{parsing_nossaflex}},
\code{\link{parsing_custom}} or \code{\link{parsing_json}}.}

\item{row_indices}{An integer or integer vector of row numbers where new
rows should be inserted (before the row at that position).}

\item{new_rows}{A data.frame or data.table with column names matching names
in `metadata` and values to insert, e.g. Roll_Name and Camera_Brand.
Must have the same number of rows as `row_indices`.}

\item{extrapolate_data}{Logical, FALSE by default. If TRUE, the function
computes average Date_Time_Original, Latitude and Longitude values. Hence
works only with JSON data.}
}
\description{
Insert rows in metadata data.tables
}
\examples{
\dontrun{
# NOSSAFLEX data
metadata <- reading_nossaflex("tests/testthat/testdata/nossaflex_filenames.txt") |>
    parsing_nossaflex()
new_rows <- data.frame(Roll_Name = "est roll", NO = "02", SS = "125",
                       A = "1.4", FL = "50", EX = "0")
insert_missing_records(metadata, row_indices = 2L, new_rows = new_rows)

# JSON data
metadata <- parsing_json("tests/testthat/testdata/nossaflex_Vito_70_test-week.txt")
new_rows <- data.frame(
  Roll_Name    = rep("Test-week", 3),
  Roll_Number  = rep("005", 3),
  Camera_Brand = rep("Voigtlaender", 3),
  Camera_Model = rep("Vito 70", 3),
  NO = 5:7,
  SS = rep("auto", 3),
  A  = rep("auto", 3)
)
insert_missing_records(metadata, row_indices = 5:7, new_rows = new_rows,
                       extrapolate_data = TRUE)
}
}
