% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{reading_nossaflex}
\alias{reading_nossaflex}
\title{Reading nossaflex file names}
\usage{
reading_nossaflex(path)
}
\arguments{
\item{path}{if null, reading from clipboard}
}
\value{
a character vector of file names without extension
}
\description{
Reading nossaflex file names
}
\examples{
path <- system.file("extdata", "nossaflex_example.txt", package = "APMD")
reading_nossaflex(path = path)
}
