% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming.R
\name{renaming_nossaflex}
\alias{renaming_nossaflex}
\title{Batch renaming picture files with the names provided by the `analog` app}
\usage{
renaming_nossaflex(files, filenames, copy = FALSE)
}
\arguments{
\item{files}{a character vector of paths with complete file names like what
\code{\link{list.files}} when used with `full.names = TRUE` or
\code{\link[tools:fileutils]{tools::list_files_with_exts}} gives}

\item{filenames}{a character vector of target file names provided by
\code{\link{reading_nossaflex}}. With or without file extensions.}

\item{copy}{logical, FALSE by default. If TRUE, a `_copy` backup of each
file is made before renaming.}
}
\value{
A named logical vector (from \code{\link[base]{file.rename}}),
  named by the new file paths.
}
\description{
By default, the `filenames` provided by `analog` do not have file extensions.
   If the `filenames` do not have an extension, the extension of `files` is
   used. If `files` do not have an extension, an error is thrown.

   By default, the `filenames` provided by `analog` do not have path.
   If the `filenames` do not have a complete path, the path of `files` is
   used. If `files` do not have a path, an error is thrown as they cannot be found.
}
\examples{
\dontrun{
files <- tools::list_files_with_exts(
  dir = system.file("extdata", package = "APMD"),
  exts = "jpg", full.names = TRUE
)
filenames <- c("NO01_SS1s_A5.6_FL35_EX-2", "NO02_SS125_A5.6_FL35_EX+1")
renaming_nossaflex(files = files, filenames = filenames, copy = TRUE)
}
}
