% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing_exif.R
\name{editing_exif}
\alias{editing_exif}
\title{Batch editing of exif data}
\usage{
editing_exif(
  metadata,
  files,
  extra_tags = NULL,
  overwrite_original = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{metadata}{a data.frame as provided by \code{\link{parsing_nossaflex}},
\code{\link{parsing_json}} or \code{\link{parsing_frames}}.}

\item{files}{a character vector of paths with complete file names like what
\code{\link{list.files}} when used with `full.names = TRUE` or
\code{\link[tools:fileutils]{tools::list_files_with_exts}} gives}

\item{extra_tags}{A named character vector of additional exif tags to write
for every file, e.g. `c(Artist = "Jane Smith", Copyright = "2024 Jane Smith")`.
These are appended to the per-shot metadata tags.}

\item{overwrite_original}{logical, if TRUE, no copy is created by `exiftool`. See <https://exiftool.org/forum/index.php?topic=13191.msg71304#msg71304>}

\item{verbose}{passes `-v2` argument to `exiftool`}
}
\value{
Called for its side effect of writing EXIF tags to `files`.
  Returns `NULL` invisibly.
}
\description{
Batch editing of exif data
}
\details{
Editing the `maker`, `model` and various `makerNotes` tags before or during darktable editing will most likely render the file unsuable for darktable since it uses these fields for parameterising the editing treatments.
}
\examples{
\dontrun{
files <- tools::list_files_with_exts(
  dir = system.file("extdata", package = "APMD"),
  exts = "jpg", full.names = TRUE
)
metadata <- reading_nossaflex(
  path = system.file("extdata", "nossaflex_filenames.txt", package = "APMD")
) |>
  parsing_nossaflex()
editing_exif(metadata = metadata, files = files)

metadata <- data.frame(
  NO = c(1, 2), SS = c("2s", "4000"), A = c(1.4, 2.8),
  FL = c(50, 50), EX = c("+2", "-1"),
  Northing = c("N", "N"), Easting = c("E", "E"),
  Latitude = c(54.321, 54.321), Longitude = c(12.345, 12.345),
  Date_Time_Original = c("2024-03-19 21:40:40 +0000", "2024-04-20 12:20:10 +0000"),
  Camera_Brand = c("Nikon", "Nikon"), Camera_Model = c("FA", "FA"),
  Lens_Brand = c("Nikon", "Nikon"),
  Lens_Model = c("Nikkor AF 50mm d f/1.4", "Nikkor AF 50mm d f/1.4"),
  Lens_Focal_Length = c(50, 50), Lens_Maximum_Aperture = c(1.4, 1.4)
)
editing_exif(metadata = metadata, files = files, extra_tags = c(Artist = "Jane Smith"))
}


}
